+++ ports/freertos/timext.c
@@ -11,17 +11,17 @@
 
 #define time_after(goal, chasing)	((long)(goal) - (long)(chasing) < 0)
 
-void timeout_set(unsigned long *goal, unsigned long msec)
+void timeout_set(uint32_t *goal, uint32_t msec)
 {
 	*goal = xTaskGetTickCount() + pdMS_TO_TICKS(msec);
 }
 
-bool timeout_is_expired(unsigned long goal)
+bool timeout_is_expired(uint32_t goal)
 {
 	return time_after(goal, xTaskGetTickCount());
 }
 
-void sleep_ms(unsigned long msec)
+void sleep_ms(uint32_t msec)
 {
 	vTaskDelay(pdMS_TO_TICKS(msec));
 }
