+++ tests/mocks/board.cpp
@@ -1,12 +1,12 @@
 #include "CppUTestExt/MockSupport.h"
 #include "libmcu/board.h"
 
-unsigned long board_get_time_since_boot_ms(void) {
-	return mock().actualCall(__func__).returnUnsignedIntValue();
+uint32_t board_get_time_since_boot_ms(void) {
+	return (uint32_t)mock().actualCall(__func__).returnUnsignedIntValue();
 }
 
 uint64_t board_get_time_since_boot_us(void) {
-	return mock().actualCall(__func__).returnUnsignedIntValue();
+	return (uint64_t)mock().actualCall(__func__).returnUnsignedIntValue();
 }
 
 void board_reboot(void) {
@@ -14,11 +14,11 @@ void board_reboot(void) {
 }
 
 uint32_t board_random(void) {
-	return mock().actualCall(__func__).returnUnsignedIntValue();
+	return (uint32_t)mock().actualCall(__func__).returnUnsignedIntValue();
 }
 
-unsigned long board_get_current_stack_watermark(void) {
-	return mock().actualCall(__func__).returnUnsignedIntValue();
+uint32_t board_get_current_stack_watermark(void) {
+	return (uint32_t)mock().actualCall(__func__).returnUnsignedIntValue();
 }
 
 const char *board_get_serial_number_string(void) {
