+++ tests/mocks/timext.cpp
@@ -1,16 +1,14 @@
 #include "CppUTestExt/MockSupport.h"
 #include "libmcu/timext.h"
 
-bool timeout_is_expired(unsigned long goal) {
+bool timeout_is_expired(uint32_t goal) {
 	return (bool)mock().actualCall(__func__).returnIntValue();
 }
 
-void timeout_set(unsigned long *goal, unsigned long msec) {
-	mock().actualCall(__func__)
-		.withParameter("msec", msec);
+void timeout_set(uint32_t *goal, uint32_t msec) {
+	mock().actualCall(__func__).withParameter("msec", msec);
 }
 
-void sleep_ms(unsigned long msec) {
-	mock().actualCall(__func__)
-		.withParameter("msec", msec);
+void sleep_ms(uint32_t msec) {
+	mock().actualCall(__func__).withParameter("msec", msec);
 }
