+++ tests/stubs/timext.cpp
@@ -1,6 +1,6 @@
 #include "CppUTestExt/MockSupport.h"
 #include "libmcu/timext.h"
 
-void sleep_ms(unsigned long msec) {
+void sleep_ms(uint32_t msec) {
 	(void)msec;
 }
